backend/requirements/common.txt
django~=3.0.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow